handwritingapi/requirements.txt
alipay==0.7.4
amqp==5.0.6
ansicon==1.89.0
asgiref==3.3.4
asn1crypto==0.24.0
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
backports.csv==1.0.7
billiard==3.6.4.0
blessed==1.18.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.6
cssselect==1.1.0
decorator==5.0.7
defusedxml==0.5.0
demjson==2.2.4
diff-match-patch==20181111
Django==2.2.1
django-cors-headers==3.2.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-formtools==2.1
django-guardian==2.0.0
django-import-export==1.2.0
django-jet==1.0.8
django-qiniu-storage==2.3.1
django-ranged-response==0.2.0
django-redis==4.12.1
django-reversion==3.0.4
django-simple-captcha==0.5.12
django-simpleui==2021.4.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
dnspython==1.16.0
drf-dynamic-fields==0.3.1
drf-extensions==0.7.0
et-xmlfile==1.0.1
eventlet==0.30.2
future==0.17.1
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
httplib2==0.9.2
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipython==7.22.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jinxed==1.0.1
kombu==5.0.2
lxml==4.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.1
Naked==0.1.31
numpy==1.19.1
nvidia-ml-py3==7.352.0
odfpy==1.4.0
openpyxl==2.6.2
parsel==1.6.0
parso==0.8.2
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.18
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.4
PyDispatcher==2.0.5
Pygments==2.8.1
PyHamcrest==2.0.2
pyinstall==0.1.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.1
pytz==2020.1
PyYAML==5.1.1
qcloudsms-py==0.1.4
qiniu==7.2.4
queuelib==1.5.0
raven==6.10.0
redis==3.0.1
requests==2.21.0
rsa==4.7.2
Scrapy==1.6.0
scrapy-redis==0.6.8
scrapy-redis-bloomfilter==0.7.0
service-identity==18.1.0
shellescape==3.8.1
simplepro==2.3.4
simpleui==4.0.2
six==1.15.0
SQLAlchemy==1.3.2
sqlparse==0.3.1
tablib==0.13.0
tencentcloud-sdk-python==3.0.365
traitlets==5.0.5
twisted==21.2.0
tyadmin-api-cli==0.5.1
uritemplate==3.0.1
urllib3==1.24.3
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
xlrd==1.2.0
XlsxWriter==1.3.9
xlwt==1.3.0
you-get==0.4.1432
zope.event==4.5.0
zope.interface==5.4.0
